restructure all test folders to match the library they test so it is easy to locate corresponding tests
