+++ packages/caliper-core/test/common/messages/assignedMessage.js
@@ -14,34 +14,34 @@
 
 'use strict';
 
-const ReadyMessage = require('../../../../lib/common/messages/readyMessage');
-const MessageTypes = require('../../../../lib/common/utils/constants').Messages.Types;
+const AssignedMessage = require('../../../lib/common/messages/assignedMessage');
+const MessageTypes = require('../../../lib/common/utils/constants').Messages.Types;
 
 const chai = require('chai');
 chai.should();
 
-describe('ReadyMessage', () => {
+describe('AssignedMessage', () => {
     describe('Constructor', () => {
-        it('should create an ReadyMessage instance with sender, recipients, and type', () => {
+        it('should create an AssignedMessage instance with sender, recipients, and type', () => {
             const sender = 'John';
             const recipients = ['Alice', 'Bob'];
-            const message = new ReadyMessage(sender, recipients);
+            const message = new AssignedMessage(sender, recipients);
 
-            message.should.be.an.instanceOf(ReadyMessage);
+            message.should.be.an.instanceOf(AssignedMessage);
             message.sender.should.equal(sender);
             message.recipients.should.deep.equal(recipients);
-            message.type.should.equal(MessageTypes.Ready);
+            message.type.should.equal(MessageTypes.Assigned);
         });
 
         it('should set the content of the message as an empty object', () => {
-            const message = new ReadyMessage('John', ['Alice', 'Bob']);
+            const message = new AssignedMessage('John', ['Alice', 'Bob']);
 
             message.content.should.deep.equal({});
         });
 
-        it('should parse the date argument as a Date object for the ReadyMessage', () => {
+        it('should parse the date argument as a Date object for the AssignedMessage', () => {
             const date = '2024-04-25';
-            const message = new ReadyMessage('John', ['Alice', 'Bob'], date);
+            const message = new AssignedMessage('John', ['Alice', 'Bob'], date);
 
             message.date.should.be.an.instanceOf(Date);
             message.date
@@ -51,7 +51,7 @@ describe('ReadyMessage', () => {
 
         it('should set the error correctly if passed', () => {
             const error = 'Some error';
-            const message = new ReadyMessage(
+            const message = new AssignedMessage(
                 'John',
                 ['Alice', 'Bob'],
                 undefined,
@@ -61,9 +61,9 @@ describe('ReadyMessage', () => {
             message.error.should.equal(error);
         });
 
-        it('should create an ReadyMessage with undefined date and error if not provided', () => {
-            const message = new ReadyMessage('John', ['Alice', 'Bob']);
-            
+        it('should create an AssignedMessage with undefined date and error if not provided', () => {
+            const message = new AssignedMessage('John', ['Alice', 'Bob']);
+
             chai.expect(message.date).to.be.undefined;
             chai.expect(message.error).to.be.undefined;
         });
